kknn-deprecated {kknn}	R Documentation

Deprecated Functions in Package kknn

Description

These functions are provided for compatibility with older versions of R only, and may be defunct as soon as of the next release.

Usage

simulation(formula, data, runs = 10, train = TRUE, k = 11, ...)

Arguments

formula
A formula object.

data
Matrix or data frame.

runs
Number of crossvalidation runs.

train
A logical value. If TRUE the training procedure for selecting optimal values of k and kernel is performed.

k
Number or maximal number of neighbors considered, dependent of choice for train.

...
Further arguments passed to or from other methods.

Value

A matrix, containing the mean and variance of the misclassification error, the absolute and the squared distances.

References

Hechenbichler K. and Schliep K.P. (2004) Weighted k-Nearest-Neighbor Techniques and Ordinal Classification, Discussion Paper 399, SFB 386, Ludwig-Maximilians University Munich (http://www.stat.uni-muenchen.de/sfb386/papers/dsp/paper399.ps)

See Also

Defunct and Deprecated

[Package kknn version 1.3.1 Index]