Munich Rent Standard Database (1994)

Many german cities compose so-called rent standards to make a decision making instrument available to tenants, landlords, renting advisory boards and experts. The rent standards are used in particular for the determination of the local comparative rent (i.e. net rent as a function of household size, equipment, year of construction, etc.). For the composition of the rent standards, a representative random sample is drawn from all relevant households, and the interesting data are determined by interviewers by means of questionnaires. The dataset contains the data of 1082 households interviewed for the munich rent standard 1994.

data(miete)

Format

A data frame with 1082 observations on the following 18 variables.

nm

Net rent in DM.

wfl

Floor space in sqm.

bj

Year of construction.

bad0

Bathroom in apartment?
1 : no
0 : yes

zh

Central heating?
1 : yes
0 : no

ww0

Hot water supply?
1 : no
0 : yes

badkach

Tiled bathroom?
1 : yes
0 : no

fenster

Window type:
1 : plain windows
0 : state-of-the-art windows

kueche

Kitchen type
1 : well equipped kitchen
0 : plain kitchen

mvdauer

Lease duration in years.

bjkat

Age category of the building (bj categorized)
1 : built before 1919
2 : built between 1919 and 1948
3 : built between 1949 and 1965
4 : built between 1966 and 1977
5 : built between 1978 and 1983
6 : built after 1983

wflkat

Floor space category (wfl categorized):
1 : less than 50 sqm
2 : between 51 sqm and 80 sqm
3 : at least 81 sqm

nmqm

Net rent per sqm.

rooms

Number of rooms in household.

nmkat

Net rent category (nm categorized):
1 : less than 500 DM
2 : between 500 DM and 675 DM
3 : between 675 DM and 850 DM
4 : between 850 DM and 1150 DM
5 : at least 1150 DM

adr

Address type:
1 : bad
2 : average
3 : good

wohn

Residential type:
1 : bad
2 : average
3 : good

Source

Fahrmeir, L., Kuenstler, R., Pigeot, I. und Tutz, G. (1997): Statistik: der Weg zur Datenanalyse, Springer, Berlin. http://www.stat.uni-muenchen.de/service/datenarchiv

The data were converted to R format by klaus.schliep@gmail.com.

Examples

data(miete)
str(miete)

#> 'data.frame':	1082 obs. of  17 variables:
#>  $ nm     : num  693 737 732 1295 395 ...
#>  $ wfl    : int  50 70 50 55 46 94 28 36 33 57 ...
#>  $ bj     : num  1972 1972 1972 1893 1957 ...
#>  $ bad0   : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ zh     : Factor w/ 2 levels "0","1": 2 2 2 2 1 2 2 2 2 2 ...
#>  $ ww0    : Factor w/ 2 levels "0","1": 1 1 1 1 2 1 1 1 1 1 ...
#>  $ badkach: Factor w/ 2 levels "0","1": 1 1 1 1 1 2 1 1 1 2 ...
#>  $ fenster: Factor w/ 2 levels "0","1": 1 1 1 1 1 1 2 1 1 1 ...
#>  $ kueche : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 2 2 1 2 ...
#>  $ mvdauer: int  2 26 1 0 27 2 9 3 1 9 ...
#>  $ bjkat  : Ord.factor w/ 6 levels "1"<"2"<"3"<"4"<..: 4 4 4 1 3 4 4 4 4 6 ...
#>  $ wflkat : Ord.factor w/ 3 levels "1"<"2"<"3": 1 2 1 2 1 3 1 1 1 2 ...
#>  $ nmqm   : num  13.87 10.52 14.64 23.55 8.59 ...
#>  $ rooms  : int  1 3 1 3 3 4 1 1 1 2 ...
#>  $ nmkat  : Ord.factor w/ 5 levels "1"<"2"<"3"<"4"<..: 3 3 3 5 1 5 1 3 4 2 ...
#>  $ adr    : Ord.factor w/ 3 levels "1"<"2"<"3": 2 2 2 2 2 2 2 2 2 2 ...
#>  $ wohn   : Ord.factor w/ 3 levels "1"<"2"<"3": 2 2 2 2 2 2 2 2 2 2 ...